backend/requirements.txt
csv-diff
importlib-metadata==6.11.0
prov
uuid
SPARQLWrapper
numpy
pandas
pytest
regex
Flask==2.3.0
flasgger==0.5.4
flask-oidc==2.2.0
flask-restx
flask_swagger_ui
dataclasses
flask-cors
requests==2.31.0
config
urllib3
PyJWT==2.9.0